requirements.txt
# Core RAG Framework
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0

# Vector Database
chromadb>=0.4.20

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.2
unstructured>=0.11.0

# Embeddings
sentence-transformers>=2.2.2

# Web Interface (Optional)
streamlit>=1.29.0
gradio>=4.8.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tqdm>=4.66.0
pandas>=2.1.0
numpy>=1.24.0

# Development and Testing
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0

# Monitoring and Logging
loguru>=0.7.2
rich>=13.7.0

# Security
cryptography>=41.0.0
